requirements.hf_train.txt
absl-py==2.1.0
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
-e git+https://github.com/huggingface/alignment-handbook.git@ae3f44fc7d8003d706752ca06f689574dffa3b76#egg=alignment_handbook
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==24.2.0
autocommand==2.2.2
backports.tarfile==1.2.0
bitsandbytes==0.44.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
datasets==3.1.0
deepspeed==0.15.4
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@c69939aa85460e8135f40bce908a6cddaa73065f
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
evaluate==0.4.0
filelock==3.16.1
flash_attn==2.7.0.post2
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.68.0
hf_transfer==0.1.8
hjson==3.1.0
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
iopath==0.1.9
itsdangerous==2.0.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0rc1
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
ninja==1.11.1.2
numpy==2.1.2
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
peft==0.13.2
pillow==11.0.0
platformdirs==4.3.6
portalocker==3.0.0
propcache==0.2.0
protobuf==3.20.2
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.0.0
pycocotools==2.0.8
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytorch-triton-rocm==3.1.0+cf34004b8a
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
responses==0.18.0
rich==13.9.4
safetensors==0.4.6.dev0
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
tokenizers==0.20.3
tomli==2.0.1
torch==2.6
torchaudio==2.5.0.dev20241113+rocm6.2
torchvision==0.20.0.dev20241113+rocm6.2
tqdm==4.67.1
transformers==4.46.3
trl==0.12.1
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.7
Werkzeug==3.1.3
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.0
zipp==3.19.2
wandb==0.19.10
lmeval@git+https://github.com/PolyAgent/lm-evaluation-harness